+++ test/libsolidity/syntaxTests/types/mapping/argument_internal.sol
@@ -1,0 +1,7 @@
+// This is expected to fail now, but may work in the future.
+contract C {
+    function f(mapping(uint => uint) storage) internal pure {
+    }
+}
+// ----
+// TypeError: (89-110): Type is required to live outside storage.