+++ test/libsolidity/syntaxTests/types/mapping/library_argument_public.sol
@@ -1,0 +1,6 @@
+library L {
+    function f(mapping(uint => uint) storage) public pure {
+    }
+}
+// ----
+// TypeError: (27-48): Type is required to live outside storage.