+++ test/libsolidity/syntaxTests/types/mapping/library_return_external.sol
@@ -1,0 +1,10 @@
+library L
+{
+    function f(mapping(uint => uint) storage a, mapping(uint => uint) storage b, bool c) external pure returns(mapping(uint => uint) storage) {
+        return c ? a : b;
+    }
+}
+// ----
+// TypeError: (27-58): Type is required to live outside storage.
+// TypeError: (60-91): Type is required to live outside storage.
+// TypeError: (123-144): Type is required to live outside storage.