+++ test/libsolidity/syntaxTests/types/mapping/library_return_private.sol
@@ -1,0 +1,6 @@
+library L
+{
+    function f(mapping(uint => uint) storage a, mapping(uint => uint) storage b, bool c) private pure returns(mapping(uint => uint) storage) {
+        return c ? a : b;
+    }
+}